+++ client/src/components/dashboard/Dashboard.js
@@ -16,38 +16,29 @@ import {
     Feed
 } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
-import React, { useEffect, useState, useRef } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React, { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
 import axios from 'axios';
 import date from 'date-and-time';
 import ordinal from 'date-and-time/plugin/ordinal';
 import queryString from 'query-string';
-import { truncateString, capitalizeFirstLetter, isEmptyString } from '../util/HelperFunctions.js';
 
+import {
+    truncateString,
+    capitalizeFirstLetter,
+    isEmptyString
+} from '../util/HelperFunctions.js';
 import useGlobalError from '../error/ErrorHooks.js';
 
 const Dashboard = (props) => {
 
-    const dispatch = useDispatch();
     const { handleGlobalError } = useGlobalError();
     const user = useSelector((state) => state.user);
     const org = useSelector((state) => state.org);
 
-    const emptyAnnouncement = {
-        title: "",
-        message: "",
-        org: "",
-        author: {
-            firstName: "",
-            lastName: "",
-            avatar: "",
-            uuid: ""
-        }
-    };
-
     /* Data */
     const [announcements, setAnnouncements] = useState([]);
-    const [recentAnnouncement, setRecentAnnouncement] = useState(emptyAnnouncement);
+    const [recentAnnouncement, setRecentAnnouncement] = useState({});
     const [projects, setProjects] = useState([]);
     const [recentProjects, setRecentProjects] = useState([]);
 
@@ -58,7 +49,6 @@ const Dashboard = (props) => {
     const [loadedAllProjects, setLoadedAllProjects] = useState(false);
     const [loadedRecentProjects, setLoadedRecentProjects] = useState(false);
     const [hasRecentAnnouncement, setHasRecentAnnouncement] = useState(false);
-    const [hasRecentProjects, setHasRecentProjects] = useState(false);
     const [showNASuccess, setShowNASuccess] = useState(false);
 
     // New Announcement Modal
@@ -71,7 +61,7 @@ const Dashboard = (props) => {
 
     // Announcement View Modal
     const [showAVModal, setShowAVModal] = useState(false);
-    const [avAnnouncement, setAVAnnouncement] = useState(emptyAnnouncement);
+    const [avAnnouncement, setAVAnnouncement] = useState({});
 
     // New Member Modal
     const [showNMModal, setShowNMModal] = useState(false);
@@ -131,6 +121,7 @@ const Dashboard = (props) => {
         axios.get('/announcements/recent').then((res) => {
             if (!res.data.err) {
                 if (res.data.announcement !== null) {
+                    console.log(res.data.announcement);
                     const { date, time } = parseDateAndTime(res.data.announcement.createdAt);
                     setRecentAnnouncement({
                         ...res.data.announcement,
@@ -142,11 +133,12 @@ const Dashboard = (props) => {
                     setHasRecentAnnouncement(false);
                 }
             } else {
-                console.log(res.data.errMsg);
+                handleGlobalError(res.data.errMsg);
             }
             setLoadedRecentAnnouncement(true);
         }).catch((err) => {
             handleGlobalError(err);
+            setLoadedRecentAnnouncement(true);
         });
     };
 
@@ -154,15 +146,19 @@ const Dashboard = (props) => {
      * Load the most recent project via GET
      * request and update the UI accordingly.
      */
-    const getRecentProjects = () => { // TODO: update to new Projects paradigm
+    const getRecentProjects = () => {
         axios.get('/projects/recent').then((res) => {
             if (!res.data.err) {
+                if (res.data.projects && Array.isArray(res.data.projects)) {
+                    setRecentProjects(res.data.projects);
+                }
             } else {
-                console.log(res.data.errMsg);
+                handleGlobalError(res.data.errMsg);
             }
             setLoadedRecentProjects(true);
         }).catch((err) => {
             handleGlobalError(err);
+            setLoadedRecentProjects(true);
         });
     };
 
@@ -173,7 +169,8 @@ const Dashboard = (props) => {
     const getAnnouncements = () => {
         axios.get('/announcements/all').then((res) => {
             if (!res.data.err) {
-                if (res.data.announcements !== null) {
+                if (res.data.announcements && Array.isArray(res.data.announcements)) {
+                    console.log(res.data.announcements);
                     var announcementsForState = [];
                     res.data.announcements.forEach((item) => {
                         const { date, time } = parseDateAndTime(item.createdAt);
@@ -193,11 +190,12 @@ const Dashboard = (props) => {
                     setAnnouncements(announcementsForState);
                 }
             } else {
-                console.log(res.data.err);
+                handleGlobalError(res.data.errMsg);
             }
             setLoadedAllAnnouncements(true);
         }).catch((err) => {
             handleGlobalError(err);
+            setLoadedAllAnnouncements(true);
         });
     };
 
@@ -205,16 +203,32 @@ const Dashboard = (props) => {
      * Load the user's active projects via GET
      * request and update the UI accordingly.
      */
-    const getProjects = () => { // TODO: update to new Projects paradigm
+    const getProjects = () => {
         axios.get('/projects/all').then((res) => {
             if (!res.data.err) {
+                if (res.data.projects && Array.isArray(res.data.projects)) {
+                    setProjects(res.data.projects);
+                }
             } else {
-                console.log(res.data.errMsg);
+                handleGlobalError(res.data.errMsg);
             }
             setLoadedAllProjects(true);
         }).catch((err) => {
             handleGlobalError(err);
+            setLoadedAllProjects(true);
         });
+    };
+
+    /**
+     * Open the New Announcement modal and ensure
+     * all form fields are reset to their defaults.
+     */
+    const openNAModal = () => {
+        resetNAForm();
+        setNATitle('');
+        setNAMessage('');
+        setNAGlobal(false);
+        setShowNAModal(true);
     };
 
     /**
@@ -264,7 +278,7 @@ const Dashboard = (props) => {
     const postNewAnnouncement = () => {
         resetNAForm();
         if (validateNAForm()) {
-            axios.post('/announcements/create', {
+            axios.post('/announcement', {
                 title: naTitle,
                 message: naMessage,
                 global: naGlobal
@@ -301,7 +315,7 @@ const Dashboard = (props) => {
      */
     const closeAVModal = () => {
         setShowAVModal(false);
-        setAVAnnouncement(emptyAnnouncement);
+        setAVAnnouncement({});
     };
 
     /**
@@ -317,11 +331,14 @@ const Dashboard = (props) => {
                         <Grid verticalAlign='middle'>
                             <Grid.Row columns={2}>
                                 <Grid.Column>
-                                    <Header as='h2' className='announcements-header'>Announcements <span className='gray-span'>(50 most recent)</span></Header>
+                                    <Header as='h2' className='announcements-header'>
+                                        Announcements <span className='gray-span'>(50 most recent)</span>
+                                    </Header>
                                 </Grid.Column>
                                 <Grid.Column>
-                                    {(user.isCampusAdmin || user.isSuperAdmin) &&
-                                        <Button color='green' floated='right' onClick={() => { setShowNAModal(true) }}>
+                                    {((user.hasOwnProperty('isCampusAdmin') && user.isCampusAdmin === true) ||
+                                        (user.hasOwnProperty('isCampusAdmin') && user.isSuperAdmin === true)) &&
+                                        <Button color='green' floated='right' onClick={openNAModal}>
                                             <Icon name='add' />
                                             New
                                         </Button>
@@ -332,7 +349,7 @@ const Dashboard = (props) => {
                         <Divider />
                         {showNASuccess &&
                             <Message
-                                onDismiss={() => { setShowNASuccess(false) }}
+                                onDismiss={() => setShowNASuccess(false)}
                                 header='Announcement Successfully Posted!'
                                 icon='check circle outline'
                                 positive
@@ -346,7 +363,7 @@ const Dashboard = (props) => {
                             <Feed>
                             {announcements.map((item, index) => {
                                 return (
-                                    <Feed.Event key={index} onClick={() => { openAVModal(index)}} className='announcement'>
+                                    <Feed.Event key={index} onClick={() => openAVModal(index)} className='announcement'>
                                         <Feed.Label image={`${item.author.avatar}`} />
                                         <Feed.Content className='announcement-content'>
                                             <Feed.Summary>
@@ -375,25 +392,32 @@ const Dashboard = (props) => {
                             loading={!loadedAllProjects}
                         >
                             <List
-                                relaxed
                                 divided
+                                size='large'
+                                relaxed='very'
                             >
-                            {projects.map((item, index) => {
-                                const itemDate = new Date(item.lastUpdate.createdAt);
-                                item.updatedDate = date.format(itemDate, 'MMM DDD, YYYY');
-                                item.updatedTime = date.format(itemDate, 'h:mm A');
-                                return (
-                                    <List.Item key={index} className='dashboard-list-item'>
-                                        <List.Icon name='book' size='large' verticalAlign='middle' />
-                                        <List.Content>
-                                            <List.Header className='recent-announcement-title' as={Link} to={`/harvesting/projects/${item.projectID}`}>{item.title}</List.Header>
-                                            <List.Description>
-                                                <span className='author-info-span gray-span'>Last updated on {item.updatedDate} at {item.updatedTime}</span>
-                                            </List.Description>
-                                        </List.Content>
-                                    </List.Item>
-                                );
-                            })}
+                            {(projects.length > 0) &&
+                                projects.map((item) => {
+                                    const itemDate = new Date(item.updatedAt);
+                                    item.updatedDate = date.format(itemDate, 'MMM DDD, YYYY');
+                                    item.updatedTime = date.format(itemDate, 'h:mm A');
+                                    return (
+                                        <List.Item key={item.projectID} className='dashboard-list-item'>
+                                            <List.Icon name='folder' size='large' verticalAlign='middle' />
+                                            <List.Content>
+                                                <List.Header as={Link} to={`/projects/${item.projectID}`}>
+                                                    {item.title}
+                                                </List.Header>
+                                                <List.Description>
+                                                    <span className='project-meta gray-span'>
+                                                        Last updated on {item.updatedDate} at {item.updatedTime}
+                                                    </span>
+                                                </List.Description>
+                                            </List.Content>
+                                        </List.Item>
+                                    );
+                                })
+                            }
                             {(projects.length === 0) &&
                                 <p>You don't have any projects right now.</p>
                             }
@@ -407,17 +431,19 @@ const Dashboard = (props) => {
                         <Header as='h2'>Home</Header>
                         <Divider />
                         <Segment basic>
-                            <Header size='medium' onClick={() => { setCurrentView('announcements')}} className='pointer-hover'>Announcements <span className='gray-span'>(most recent)</span></Header>
+                            <Header size='medium' onClick={() => setCurrentView('announcements')} className='pointer-hover'>
+                                Announcements <span className='gray-span'>(most recent)</span>
+                            </Header>
                             {!loadedRecentAnnouncement &&
                                 <p><em>Loading...</em></p>
                             }
                             {(loadedRecentAnnouncement && hasRecentAnnouncement) &&
                                 <Message icon>
-                                    <Icon name='announcement' />
-                                    <Message.Content>
+                                    <Icon name='announcement' circular fitted className='recent-announcement-icon' />
+                                    <Message.Content className='recent-announcement-content'>
                                         <Message.Header className="recent-announcement-title">{recentAnnouncement.title}</Message.Header>
-                                        {recentAnnouncement.message}<br />
-                                        <span className='author-info-span gray-span'>by {recentAnnouncement.author.firstName} {recentAnnouncement.author.lastName} on {recentAnnouncement.date} at {recentAnnouncement.time}</span>
+                                        <p>{truncateString(recentAnnouncement.message, 250)}</p>
+                                        <p className='recent-announcement-meta gray-span'>by {recentAnnouncement.author?.firstName} {recentAnnouncement.author?.lastName} on {recentAnnouncement.date} at {recentAnnouncement.time}</p>
                                     </Message.Content>
                                 </Message>
                             }
@@ -427,28 +453,36 @@ const Dashboard = (props) => {
                         </Segment>
                         <Divider />
                         <Segment basic>
-                            <Header size='medium' onClick={() => { setCurrentView('projects') }} className='pointer-hover'>Your Projects <span className='gray-span'>(overview)</span></Header>
-                            <List divided size='large'>
+                            <Header size='medium' onClick={() => setCurrentView('projects')} className='pointer-hover'>
+                                Your Projects <span className='gray-span'>(overview)</span>
+                            </Header>
+                            <List divided size='large' relaxed='very'>
                                 {!loadedRecentProjects &&
                                     <p><em>Loading...</em></p>
                                 }
-                                {(loadedRecentProjects && hasRecentProjects) &&
-                                    projects.map((item, index) => {
-                                        const itemDate = new Date(item.lastUpdate.createdAt);
+                                {(loadedRecentProjects && recentProjects.length > 0) &&
+                                    recentProjects.map((item) => {
+                                        const itemDate = new Date(item.updatedAt);
                                         item.updatedDate = date.format(itemDate, 'MMM DDD, YYYY');
                                         item.updatedTime = date.format(itemDate, 'h:mm A');
                                         return (
-                                            <List.Item>
-                                              <List.Icon name='book' size='large' verticalAlign='middle' />
+                                            <List.Item key={item.projectID}>
+                                              <List.Icon name='folder' size='large' verticalAlign='middle' />
                                               <List.Content>
-                                                <List.Header as={Link} to={`/projects/${item.projectID}`}>{item.title}</List.Header>
-                                                <List.Description as='a'>Last updated on {item.updatedDate} at {item.updatedTime}</List.Description>
+                                                <List.Header as={Link} to={`/projects/${item.projectID}`}>
+                                                    {item.title}
+                                                </List.Header>
+                                                <List.Description>
+                                                    <span className='project-meta gray-span'>
+                                                        Last updated on {item.updatedDate} at {item.updatedTime}
+                                                    </span>
+                                                </List.Description>
                                               </List.Content>
                                             </List.Item>
                                         );
                                     })
                                 }
-                                {(loadedRecentProjects && !hasRecentProjects) &&
+                                {(loadedRecentProjects && recentProjects.length === 0) &&
                                     <p>You have no recent projects right now.</p>
                                 }
                             </List>
@@ -476,113 +510,54 @@ const Dashboard = (props) => {
                                 {user.firstName}
                             </Header>
                         </Menu.Item>
-                      <Menu.Item
-                        name='home'
-                        onClick={() => { setCurrentView('home') }}
-                        active={currentView === 'home'}
-                        color={currentView === 'home' ? 'blue' : 'black'}
-                      >
-                        Home
-                      </Menu.Item>
-                      <Menu.Item
-                        name='announcements'
-                        onClick={() => { setCurrentView('announcements') }}
-                        active={currentView === 'announcements'}
-                        color={currentView === 'announcements' ? 'blue' : 'black'}
-                      >
-                        Announcements
-                      </Menu.Item>
-                      <Menu.Item
-                        name='projects'
-                        onClick={() => { setCurrentView('projects') }}
-                        active={currentView === 'projects'}
-                        color={currentView === 'projects' ? 'blue' : 'black'}
+                        <Menu.Item
+                            name='home'
+                            onClick={() => setCurrentView('home')}
+                            active={currentView === 'home'}
+                            color={currentView === 'home' ? 'blue' : 'black'}
+                        >
+                            Home
+                        </Menu.Item>
+                        <Menu.Item
+                            name='announcements'
+                            onClick={() => setCurrentView('announcements')}
+                            active={currentView === 'announcements'}
+                            color={currentView === 'announcements' ? 'blue' : 'black'}
+                        >
+                            Announcements
+                        </Menu.Item>
+                        <Menu.Item
+                            name='projects'
+                            onClick={() => setCurrentView('projects')}
+                            active={currentView === 'projects'}
+                            color={currentView === 'projects' ? 'blue' : 'black'}
                         >
                             Your Projects
                         </Menu.Item>
                         <Menu.Item>&nbsp;
                         </Menu.Item>
-                      <Menu.Item href='https://libretexts.org' target='_blank' rel='noopener noreferrer'>
-                        LibreTexts.org
-                        <Icon name='external' />
-                      </Menu.Item>
+                        {((user.hasOwnProperty('isSuperAdmin') && user.isSuperAdmin === true) ||
+                            (user.hasOwnProperty('isSuperAdmin') && user.isCampusAdmin === true)) &&
+                            <Menu.Item
+                                as={Link}
+                                to='/controlpanel'
+                            >
+                                Control Panel
+                                <Icon name='dashboard' />
+                            </Menu.Item>
+                        }
+                        <Menu.Item
+                            href='https://libretexts.org'
+                            target='_blank'
+                            rel='noopener noreferrer'
+                        >
+                            LibreTexts.org
+                            <Icon name='external' />
+                        </Menu.Item>
                     </Menu>
                 </Grid.Column>
-                <Grid.Column width={9}>
+                <Grid.Column width={13}>
                     <View />
-                </Grid.Column>
-                <Grid.Column width={4}>
-                    <Segment>
-                        <Header as='h3'>Libraries</Header>
-                        <Divider />
-                        <Menu vertical fluid secondary size='tiny'>
-                          <Menu.Item href='https://bio.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='dna' /></span>
-                            Biology
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://biz.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='dollar' /></span>
-                            Business
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://chem.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='flask' /></span>
-                            Chemistry
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://eng.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='wrench' /></span>
-                            Engineering
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://espanol.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='language' /></span>
-                            Español
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://geo.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='globe' /></span>
-                            Geosciences
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://human.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='address book' /></span>
-                            Humanities
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://math.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='subscript' /></span>
-                            Mathematics
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://med.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='first aid' /></span>
-                            Medicine
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://phys.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='rocket' /></span>
-                            Physics
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://socialsci.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='users' /></span>
-                            Social Science
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://stats.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='chart pie' /></span>
-                            Statistics
-                            <Icon name='external' />
-                          </Menu.Item>
-                          <Menu.Item href='https://workforce.libretexts.org/' target='_blank' rel='noopener noreferrer'>
-                            <span><Icon name='briefcase' /></span>
-                            Workforce
-                            <Icon name='external' />
-                          </Menu.Item>
-                        </Menu>
-                    </Segment>
                 </Grid.Column>
             </Grid.Row>
             <Modal
@@ -602,7 +577,7 @@ const Dashboard = (props) => {
                                 type='text'
                                 placeholder='Enter title...'
 
-                                onChange={(e) => { setNATitle(e.target.value) }}
+                                onChange={(e) => setNATitle(e.target.value)}
                                 value={naTitle}
                             />
                         </Form.Field>
@@ -613,17 +588,17 @@ const Dashboard = (props) => {
                             <label>Message</label>
                             <Form.TextArea
                                 placeholder='Enter message...'
-                                onChange={(e) => { setNAMessage(e.target.value) }}
+                                onChange={(e) => setNAMessage(e.target.value)}
                                 value={naMessage}
                             />
                         </Form.Field>
-                        {user.isSuperAdmin &&
+                        {(user.hasOwnProperty('isSuperAdmin') && user.isSuperAdmin === true) &&
                             (
                                 <div className='mb-2p'>
                                     <p><strong><em>Super Administrator Options</em></strong> <span className='muted-text'>(use caution)</span></p>
                                     <Form.Field>
                                         <Form.Checkbox
-                                            onChange={() => { setNAGlobal(!naGlobal) }}
+                                            onChange={() => setNAGlobal(!naGlobal)}
                                             checked={naGlobal}
                                             label="Send globally"
                                         />
@@ -642,7 +617,11 @@ const Dashboard = (props) => {
                     </span>
                 </Modal.Content>
                 <Modal.Actions>
-                    <Button onClick={closeNAModal}>Cancel</Button>
+                    <Button
+                        onClick={closeNAModal}
+                    >
+                        Cancel
+                    </Button>
                     <Button
                         color='green'
                         onClick={postNewAnnouncement}
@@ -664,19 +643,36 @@ const Dashboard = (props) => {
                 </Modal.Header>
                 <Modal.Content>
                     <Header as='h4'>
-                        <Image avatar src={`${avAnnouncement.author.avatar}`} />
+                        <Image avatar src={`${avAnnouncement.author?.avatar || '/mini_logo.png'}`} />
                         <Header.Content>
-                            {avAnnouncement.author.firstName} {avAnnouncement.author.lastName}
+                            {avAnnouncement.author?.firstName} {avAnnouncement.author?.lastName}
                             <Header.Subheader>
                                 {avAnnouncement.date} at {avAnnouncement.time}
                             </Header.Subheader>
                         </Header.Content>
                     </Header>
                     <Modal.Description className='announcement-view-text'>{avAnnouncement.message}</Modal.Description>
-                    <span className='gray-span'>Sent to: {capitalizeFirstLetter(avAnnouncement.org)}</span>
+                    <span className='gray-span'>Sent to: {capitalizeFirstLetter(avAnnouncement.org?.shortName || 'Unknown')}</span>
+                    {(user.hasOwnProperty('uuid') && avAnnouncement.hasOwnProperty('author') &&
+                        avAnnouncement.author.hasOwnProperty('uuid') && avAnnouncement.author.uuid === user.uuid) &&
+                        <Button
+                            color='red'
+                            className='float-right'
+                            fluid
+                            disabled
+                        >
+                            <Icon name='trash' />
+                            Delete
+                        </Button>
+                    }
                 </Modal.Content>
                 <Modal.Actions>
-                    <Button onClick={closeAVModal} color='blue'>Done</Button>
+                    <Button
+                        onClick={closeAVModal}
+                        color='blue'
+                    >
+                        Done
+                    </Button>
                 </Modal.Actions>
             </Modal>
             {/* New Member Modal */}
@@ -695,9 +691,7 @@ const Dashboard = (props) => {
                 </Modal.Content>
                 <Modal.Actions>
                     <Button
-                        onClick={() => {
-                            setShowNMModal(false);
-                        }}
+                        onClick={() => setShowNMModal(false)}
                         color='blue'
                     >
                         Done